engine: Add Eval module for evaluating positions
This change introduces an evaluation function for further minmax
searching. It takes into account raw material, isolated pawns, doubled
pawns and basic piece mobility.

+++ engine/Engine/BoardState.hs
@@ -24,7 +24,7 @@ data BoardState = BoardState {
     castling :: Int, -- k q K Q
     sideToMove :: Color,
     ply :: Int,
-    eval :: Int
+    eval :: Float
 }
 
 -- Board helper functions
